requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
Django==5.1.5
django-crispy-forms==2.3
django-paypal==2.1
idna==3.10
pillow==11.1.0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
